pageflip

+++ libs/micro_serde/src/serde_ron.rs
@@ -603,7 +603,7 @@ macro_rules!impl_ser_de_ron_float {
     }
 }
 
-impl_ser_de_ron_unsigned!(usize, std::u64::MAX);
+impl_ser_de_ron_unsigned!(usize, std::usize::MAX);
 impl_ser_de_ron_unsigned!(u64, std::u64::MAX);
 impl_ser_de_ron_unsigned!(u32, std::u32::MAX);
 impl_ser_de_ron_unsigned!(u16, std::u16::MAX);
